+++ Models/DamageThree.cs
@@ -2,11 +2,11 @@
 {
     public class DamageThree
     {
-        public int? RepairAndReplaceExpense { get; set; }
-        public int? PaintingLabourExpense { get; set; }
-        public int? NewBodyPartsExpense { get; set; }
+        public double? RepairAndReplaceExpense { get; set; }
+        public double? PaintingLabourExpense { get; set; }
+        public double? NewBodyPartsExpense { get; set; }
 
-        public DamageThree(int param1, int param2, int param3)
+        public DamageThree(double param1, double param2, double param3)
         {
             this.RepairAndReplaceExpense = param1;
             this.PaintingLabourExpense = param2;
